- removed some "using std::cerr" etc from .h or .inl files to avoid conflicts. This then needed changes in some files that include these files - a few extra explicit float or int casts to reduce number of warnings by VC++ - replaced a few output lines to std::cerr with calls to stir::info

+++ src/include/stir/MultipleProjData.h
@@ -107,7 +107,7 @@ public:
   /*! \deprecated */
   unsigned int get_num_gates() const
   {
-    return _proj_datas.size();
+    return static_cast<unsigned int>(_proj_datas.size());
   }
 
 protected:
